Port output unit test to test_HIP_RA.py
